+++ src/pages/Profile.tsx
@@ -1,8 +1,9 @@
 import { useState, useRef, useEffect } from 'react';
-import { Pencil, Check, Save, X, AlertCircle } from 'lucide-react';
+import { Pencil, Save, X } from 'lucide-react';
 import { useUser } from '@/context/UserContext';
 import ActionButton from '@/components/our-components/actionButton';
-import { validateField } from '@/utils/formValidator';
+import ProfileField from '@/components/our-components/profileField';
+import { profileValidator } from '@/utils/profileValidator';
 
 interface SkillOption {
   label: string;
@@ -34,7 +35,6 @@ export default function ProfilePage() {
     Record<string, number>
   >({});
   const fileInputRef = useRef<HTMLInputElement>(null);
-  const inputRefs = useRef<Record<string, HTMLInputElement | null>>({});
 
   useEffect(() => {
     const hasUnsavedChanges =
@@ -87,7 +87,7 @@ export default function ProfilePage() {
     // If already editing another field, validate current field first
     if (editingField && editingField !== field) {
       const currentValue = tempValues[editingField] || '';
-      const validation = validateField(editingField, currentValue);
+      const validation = profileValidator(editingField, currentValue);
 
       if (!validation.isValid) {
         setValidationErrors((prev) => ({
@@ -126,7 +126,7 @@ export default function ProfilePage() {
         return;
       }
 
-      const validation = validateField(field, currentValue);
+      const validation = profileValidator(field, currentValue);
 
       if (!validation.isValid) {
         setValidationErrors((prev) => ({
@@ -167,210 +167,12 @@ export default function ProfilePage() {
     }
   };
 
-  // Props type for the inner component
-  interface EditableFieldProps {
-    label: string;
-    field: string;
-  }
-
-  const EditableField = ({ label, field }: EditableFieldProps) => {
-    useEffect(() => {
-      if (
-        field !== 'skills' &&
-        editingField === field &&
-        inputRefs.current[field] &&
-        cursorPositions[field] !== undefined
-      ) {
-        const input = inputRefs.current[field];
-        input?.focus();
-        input?.setSelectionRange(
-          cursorPositions[field],
-          cursorPositions[field],
-        );
-      }
-    }, [field, editingField, cursorPositions]);
-
-    return (
-      <div>
-        <span className="font-semibold flex items-center">
-          {label}
-          {editingField === field ? (
-            <Check
-              onClick={() => saveEditing(field)}
-              className="w-4 h-4 ml-2 text-green-600 cursor-pointer hover:text-green-700"
-            />
-          ) : (
-            <Pencil
-              onClick={() => startEditing(field)}
-              className="w-4 h-4 ml-2 text-gray-500 cursor-pointer hover:text-gray-700"
-            />
-          )}
-        </span>
-
-        {editingField === field ? (
-          <div className="flex flex-col gap-2 mt-1">
-            {field === 'skills' ? (
-              <div className="grid grid-cols-2 gap-2">
-                {skillOptions.map((skillOption) => {
-                  const currentSkills = tempValues['skills']
-                    ? tempValues['skills']
-                        .split(', ')
-                        .filter((s) => s.trim() !== '')
-                    : user.providerProfile?.skills || [];
-
-                  const isChecked = currentSkills.includes(skillOption.value);
-
-                  return (
-                    <label
-                      key={skillOption.value}
-                      className="flex items-center space-x-2"
-                    >
-                      <input
-                        type="checkbox"
-                        checked={isChecked}
-                        onChange={(e) => {
-                          setSkillsChanged(true);
-                          let newSkills;
-                          if (e.target.checked) {
-                            newSkills = [...currentSkills, skillOption.value];
-                          } else {
-                            newSkills = currentSkills.filter(
-                              (s) => s !== skillOption.value,
-                            );
-                          }
-
-                          // Sort skills according to the order defined in skillOptions
-                          const sortedSkills = skillOptions
-                            .filter((skill) => newSkills.includes(skill.value))
-                            .sort((a, b) => a.order - b.order)
-                            .map((skill) => skill.value);
-
-                          setTempValues((prev) => ({
-                            ...prev,
-                            skills: sortedSkills.join(', '),
-                          }));
-                        }}
-                        className="rounded"
-                      />
-                      <span className="text-sm">{skillOption.label}</span>
-                    </label>
-                  );
-                })}
-              </div>
-            ) : (
-              <>
-                <input
-                  ref={(el) => {
-                    inputRefs.current[field] = el;
-                  }}
-                  type="text"
-                  value={tempValues[field] || ''}
-                  onChange={(e) => {
-                    const newValue = e.target.value;
-                    const cursorPos = e.target.selectionStart || 0;
-                    setTempValues((prev) => ({ ...prev, [field]: newValue }));
-                    setCursorPositions((prev) => ({
-                      ...prev,
-                      [field]: cursorPos,
-                    }));
-
-                    if (validationErrors[field]) {
-                      setValidationErrors((prev) => {
-                        const newErrors = { ...prev };
-                        delete newErrors[field];
-                        return newErrors;
-                      });
-                    }
-                  }}
-                  onKeyUp={(e) => {
-                    const target = e.target as HTMLInputElement;
-                    setCursorPositions((prev) => ({
-                      ...prev,
-                      [field]: target.selectionStart || 0,
-                    }));
-                  }}
-                  onClick={(e) => {
-                    const target = e.target as HTMLInputElement;
-                    setCursorPositions((prev) => ({
-                      ...prev,
-                      [field]: target.selectionStart || 0,
-                    }));
-                  }}
-                  className={`border rounded px-2 py-1 flex-1 focus:outline-none focus:ring-2 ${
-                    validationErrors[field]
-                      ? 'border-red-500 focus:ring-red-200'
-                      : 'border-gray-300 focus:ring-blue-200'
-                  }`}
-                />
-                {validationErrors[field] && (
-                  <div className="flex items-center gap-1 text-red-500 text-sm mt-1">
-                    <AlertCircle className="w-4 h-4" />
-                    <span>{validationErrors[field]}</span>
-                  </div>
-                )}
-              </>
-            )}
-          </div>
-        ) : (
-          <>
-            {field === 'skills' ? (
-              <div className="flex flex-wrap gap-2 mt-1">
-                {(() => {
-                  const currentSkills = tempValues['skills']
-                    ? tempValues['skills']
-                        .split(', ')
-                        .filter((s) => s.trim() !== '')
-                    : user.providerProfile?.skills || [];
-
-                  return currentSkills.length ? (
-                    currentSkills.map((skillValue, index) => {
-                      const skillOption = skillOptions.find(
-                        (s) => s.value === skillValue,
-                      );
-                      return (
-                        <span
-                          key={index}
-                          className="px-3 py-1 bg-background border border-gray-300 rounded-full text-sm text-black"
-                        >
-                          {skillOption?.label || skillValue}
-                        </span>
-                      );
-                    })
-                  ) : (
-                    <p className="text-gray-500">No skills selected</p>
-                  );
-                })()}
-              </div>
-            ) : (
-              <>
-                {getDisplayValue(field) ? (
-                  <p
-                    className={
-                      tempValues[field] !== undefined &&
-                      tempValues[field] !== getFieldValue(field)
-                        ? 'text-button-upload font-medium'
-                        : ''
-                    }
-                  >
-                    {getDisplayValue(field)}
-                  </p>
-                ) : (
-                  <p className="text-gray-500">{getPlaceholderText(field)}</p>
-                )}
-              </>
-            )}
-          </>
-        )}
-      </div>
-    );
-  };
-
   const submitAllChanges = () => {
     let hasErrors = false;
     const newValidationErrors: Record<string, string> = {};
 
     Object.entries(tempValues).forEach(([field, value]) => {
-      const validation = validateField(field, value);
+      const validation = profileValidator(field, value);
       if (!validation.isValid) {
         newValidationErrors[field] = validation.error || '';
         hasErrors = true;
@@ -428,7 +230,7 @@ export default function ProfilePage() {
 
     alert('Profile updated successfully!');
 
-    // console.log('User Data: ', updatedUser);
+    console.log('User Data: ', updatedUser);
   };
 
   const canSave = hasChanges && editingField === null;
@@ -477,11 +279,93 @@ export default function ProfilePage() {
 
         {/* Editable Fields */}
         <div className="space-y-4 w-full">
-          <EditableField label="Name" field="name" />
-          <EditableField label="Telephone" field="telNumber" />
-          <EditableField label="Email" field="email" />
-          <EditableField label="Description" field="description" />
-          <EditableField label="Skill & Experience" field="skills" />
+          <ProfileField
+            label="Name"
+            field="name"
+            editingField={editingField}
+            tempValues={tempValues}
+            validationErrors={validationErrors}
+            cursorPositions={cursorPositions}
+            getDisplayValue={getDisplayValue}
+            getFieldValue={getFieldValue}
+            getPlaceholderText={getPlaceholderText}
+            startEditing={startEditing}
+            saveEditing={saveEditing}
+            setTempValues={setTempValues}
+            setCursorPositions={setCursorPositions}
+            setValidationErrors={setValidationErrors}
+            setSkillsChanged={setSkillsChanged}
+          />
+          <ProfileField
+            label="Telephone"
+            field="telNumber"
+            editingField={editingField}
+            tempValues={tempValues}
+            validationErrors={validationErrors}
+            cursorPositions={cursorPositions}
+            getDisplayValue={getDisplayValue}
+            getFieldValue={getFieldValue}
+            getPlaceholderText={getPlaceholderText}
+            startEditing={startEditing}
+            saveEditing={saveEditing}
+            setTempValues={setTempValues}
+            setCursorPositions={setCursorPositions}
+            setValidationErrors={setValidationErrors}
+            setSkillsChanged={setSkillsChanged}
+          />
+          <ProfileField
+            label="Email"
+            field="email"
+            editingField={editingField}
+            tempValues={tempValues}
+            validationErrors={validationErrors}
+            cursorPositions={cursorPositions}
+            getDisplayValue={getDisplayValue}
+            getFieldValue={getFieldValue}
+            getPlaceholderText={getPlaceholderText}
+            startEditing={startEditing}
+            saveEditing={saveEditing}
+            setTempValues={setTempValues}
+            setCursorPositions={setCursorPositions}
+            setValidationErrors={setValidationErrors}
+            setSkillsChanged={setSkillsChanged}
+          />
+          <ProfileField
+            label="Description"
+            field="description"
+            editingField={editingField}
+            tempValues={tempValues}
+            validationErrors={validationErrors}
+            cursorPositions={cursorPositions}
+            getDisplayValue={getDisplayValue}
+            getFieldValue={getFieldValue}
+            getPlaceholderText={getPlaceholderText}
+            startEditing={startEditing}
+            saveEditing={saveEditing}
+            setTempValues={setTempValues}
+            setCursorPositions={setCursorPositions}
+            setValidationErrors={setValidationErrors}
+            setSkillsChanged={setSkillsChanged}
+          />
+          <ProfileField
+            label="Skill & Experience"
+            field="skills"
+            skillOptions={skillOptions}
+            editingField={editingField}
+            tempValues={tempValues}
+            validationErrors={validationErrors}
+            cursorPositions={cursorPositions}
+            getDisplayValue={getDisplayValue}
+            getFieldValue={getFieldValue}
+            getPlaceholderText={getPlaceholderText}
+            startEditing={startEditing}
+            saveEditing={saveEditing}
+            setTempValues={setTempValues}
+            setCursorPositions={setCursorPositions}
+            setValidationErrors={setValidationErrors}
+            setSkillsChanged={setSkillsChanged}
+            userSkills={user.providerProfile?.skills}
+          />
         </div>
 
         {/* Submit Button */}
